feat: 적과 플레이어의 데미지 처리
- Enemy 와 Player에게 IDamage 인터페이스 구현
- 적과 플레이어 충돌 처리
- 투사체와 적의 총돌 처리

+++ Game/Interface/IDamageable.h
@@ -3,5 +3,5 @@
 class IDamageable
 {
 public:
-	virtual void TakeDamage(int damage) abstract;
+	virtual void TakeDamage(float damage) abstract;
 };
